add way to expand sidebar into center of page
this commit adds a centered, focused and more readable version of the sidebar.
this came up when i was at the kitchen table with my phone and wanted to go
look through all my recipes for inspiration.

it can be toggled on by clicking the lil weird quotation marks in the top left
(»). to toggle it off, click the quotation again.